moving config files
